+++ Stateless/ToggleButton.jsx
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
-import Button from 'rebass/dist/Button';
-import ButtonOutline from 'rebass/dist/ButtonOutline';
+import Button from './Button';
+import ButtonOutline from './ButtonOutline';
 
 export default function ToggleButton(props) {
 	if (props.selected) {